Change default value for RegressionAngleRange

+++ src/dbmtester/DBMTester.vb
@@ -97,7 +97,7 @@ Namespace Vitens.DynamicBandwidthMonitor
             ElseIf Parameter.ToLower.Equals("ct") Then
               CorrelationThreshold = Convert.ToDouble(Value)
             ElseIf Parameter.ToLower.Equals("ra") Then
-              RegressionAngleRange = Convert.ToInt32(Value)
+              RegressionAngleRange = Convert.ToDouble(Value)
             ElseIf Parameter.ToLower.Equals("st") Then
               StartTimestamp = Convert.ToDateTime(Value)
             ElseIf Parameter.ToLower.Equals("et") Then
